Fahrtkosten-Generator/requirements.txt
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
